browserstealth/vision_agent/requirements.txt
selenium==4.18.1
pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
pyautogui==0.9.54
google-generativeai==0.3.2
python-dotenv==1.0.1
webdriver-manager==4.0.1
requests==2.31.0
